refactor CoursePeople by deleting it. FeelsGoodMan, closes #79
